requirements.txt
awacs
boto3
click
connexion
fauxfactory
Flask-HTTPAuth
troposphere
pyopenssl
cffi
